Rawr3.DPSDK: More improvements to options pane. Almost all relevant options can be controlled in rawr3 now!
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C37777

+++ Rawr.DPSDK/CalculationOptionsPanelDPSDK.cs
@@ -40,7 +40,7 @@ namespace Rawr.DPSDK
            
 
             cbTargetLevel.SelectedItem = calcOpts.TargetLevel.ToString();
-			tbFightLength.Value = calcOpts.FightLength;
+			tbFightLength.Value = (int)calcOpts.FightLength;
 			lblFightLengthNum.Text = tbFightLength.Value.ToString();
 
             nudTargetArmor.Value = calcOpts.BossArmor;
